InformationService/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1 